requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
tensorflow>=2.16.0
pillow>=10.0.1
numpy>=1.24.3
requests>=2.25.0 